+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.1.8] - 2022-07-24
+### Bug fixes
+- Compatibility with twsapi 10.16.01
+- Added connect-simple in the gateway
+
 ## [0.1.7] - 2022-05-14
 ### Bug fixes
 - Compatibility with twsapi 10.15.02: different signature to error class in the wrapper